checksum_test is working (apparently)
