+++ xdelta3/testing/Makefile
@@ -3,3 +3,6 @@ all:
 
 xdelta3regtest:
 	(cd .. && make xdelta3regtest)
+
+xdelta3checksum:
+	(cd .. && make xdelta3checksum)
