+++ xdelta3/testing/run_release.sh
@@ -1,0 +1,2 @@
+#!/bin/sh
+(cd .. && ./run_release.sh)
